New base class for SX126x modules. Added new SX1268 module support.
